requirements.txt
fastapi==0.82.0
deta==1.1.0
python-dotenv==0.21.0
python-multipart==0.0.5